requirements.txt
clip==1.0
diffusers==0.26.3
einops==0.8.0
gradio==4.39.0
matplotlib==3.5.2
numpy==1.22.4
omegaconf==2.3.0
Pillow==9.1.1
Pillow==10.4.0
Requests==2.32.3
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.61.2
